LPS-180450 Renaming tool
