--- modules/util/portal-tools-db-partition-virtual-instance-migration/scripts/db_partition_virtual_instance_migration.bat
@@ -1,11 +1,0 @@
-@echo off
-
-pushd "%~dp0"
-
-path %PATH%;%JAVA_HOME%\bin
-
-java -jar com.liferay.portal.tools.db.partition.virtual.instance.migration.jar %*
-
-popd
-
-@echo on
