--- modules/util/portal-tools-db-partition-virtual-instance-migration/source-formatter-suppressions.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<checkstyle>
-		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/internal/util/DatabaseUtil\.java" />
-	</checkstyle>
-	<source-check>
-		<suppress checks="StringMethodsCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/internal/util/DatabaseUtil\.java" />
-		<suppress checks="StringMethodsCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/internal/util/Validator\.java" />
-	</source-check>
-</suppressions>
