+++ modules/util/portal-tools-db-partition-virtual-instance-migrator/README.markdown
@@ -1,0 +1,20 @@
+# DB Partition Virtual Instance Migrator Tool
+This tool allows to migrate a virtual instance to a new partition in a database partitioned environment. Before starting the migration, it validates that all the needed conditions for a successful migration are met.
+
+## Requirements:
+    - MySQL
+    - Database user with DDL privileges
+
+## Usage
+    usage: Liferay Portal DB Partition Virtual Instance Migrator
+    -d,--destination-jdbc-url <arg> Set the JDBC url for the default partition in the destination database.
+    -dp,--destination-password <arg> Set the destination database user password.
+    -dsp,--destination-schema-prefix <arg> Set the schema prefix for non-default partitions in destination database.
+    -du,--destination-user <arg> Set the destination database user name.
+    -h,--help Print help message.
+    -s,--source-jdbc-url <arg> Set the JDBC url for the schema in the source database.
+    -sp,--source-password <arg> Set the source database user password.
+    -su,--source-user <arg> Set the source database user name.
+
+## Execution example
+    java -jar com.liferay.portal.tools.db.partition.virtual.instance.migrator.jar -s "jdbc:mysql://localhost:3306/lpartition_xxxxx?useUnicode=true&characterEncoding=UTF-8&useFastDateParsing=false&useTimezone=true&serverTimezone=GMT" -su sourceDatabaseUser -sp sourceDatabasePassword -d "jdbc:mysql://localhost:3306/lportal?useUnicode=true&characterEncoding=UTF-8&useFastDateParsing=false&useTimezone=true&serverTimezone=GMT" -du destinationDatabaseUser -dp destinationDatabasePassword
