+++ modules/util/portal-tools-db-partition-virtual-instance-migrator/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Tools DB Partition Virtual Instance Migrator
+Bundle-SymbolicName: com.liferay.portal.tools.db.partition.virtual.instance.migrator
+Bundle-Version: 1.0.0
+Main-Class: com.liferay.portal.tools.db.partition.virtual.instance.migrator.VirtualInstanceMigrator
